Resize button works
Slopy code but it works.
Pressing a button resizes the div conationer that holds the map based on the prev values of the width.

Next step -> create an object w/ a listner to do this at the bottom of the map so that we can do this without a button and easily.