Created Div boxes for SI
Added div boexes to orginze information in the SI div.

Next step??????

Found a feature in chrome to allow us to see what it would look line on a phone and it did not look the same as when I minimized the screen so that was disapointing.

I think that we keep producing a website for desktop frist and then work on transposing it into an app.

Functionallity > Looks