Added SI content
Added a hover over option for the Total players stat. I also added another function to get current player data from the database.

The next task will be to add in the event portion of the SI.

After that is done I will stop adding new content and I will clean and comment all of the code and change around the file names. Possible condense some of the files into one file.

Added a CSS grid box to the SI area. It was much easier that I expected.