Broke some stuff - Fized some stuff
Changing gears
I do not care about the front end any more. I just want to get the functionality there and then worry about how pretty is it later

Who cares if it is pretty if it does not work also we will needing to redo it anyways with boot strap.